image_server/requirements.txt
fastapi[standard]
Pillow
numpy